I have refactored the `csv_browser` script into modules.
The `csv_browser_v5.x.py` script was a single, large file containing all the application's logic, which made it difficult to maintain and extend.

To improve this, I reorganized the code into a `csv_browser` package with the following modules:
- `main.py`: The application's entry point.
- `app.py`: The main `CSVBrowser` application class.
- `file_browser.py`: A class to manage the file browsing panel.
- `csv_viewer.py`: A class to handle the CSV viewing panel.
- `actions.py`: A module for functions triggered by toolbar actions.
- `settings.py`: A module for loading and saving application settings.
- `utils.py`: A collection of utility functions.

This new structure separates concerns and will make the code easier for you to understand and work with. I have also removed the original script.
